+++ src/components/World/Builder/hooks/useSaveToIpfs.ts
@@ -1,7 +1,7 @@
 import { useCallback, useState, useEffect } from "react";
-import { saveSceneToIpfs } from "../../../api/ipfsSaver";
-import { SceneConfiguration } from "../../../types/scene";
-import { SceneFilesLocal } from "../../../types/shared";
+import { saveSceneToIpfs } from "../../../../api/ipfsSaver";
+import { SceneConfiguration } from "../../../../types/scene";
+import { SceneFilesLocal } from "../../../../types/shared";
 import { SceneSaveStatus } from "./useBuilder";
 
 const useSaveToIpfs = ({
