+++ src/components/World/Controls/TransformAndOrbitControls.tsx
@@ -1,8 +1,6 @@
 import { Object3D } from "three";
 import { TransformControls, OrbitControls } from "@react-three/drei";
-import { TransformMode } from "../Builder/useBuilder";
-import { useCallback } from "react";
-import { Transform } from "../../../types/elements";
+import { TransformMode } from "../Builder/hooks/useBuilder";
 
 const TransformAndOrbitControls = ({
   targetElement,
