--- src/editorDb/schema.graphql
@@ -1,16 +1,0 @@
-type Author {
-  id: Int!
-  firstName: String!
-  lastName: String!
-  posts(findTitle: String): [Post]
-}
-
-type Post {
-  id: Int!
-  title: String!
-  author: Author
-}
-
-type Query {
-  posts: [Post]
-}
