--- src/editorDb/world.graphql
@@ -1,12 +1,0 @@
-type Environment {
-  fileUrl: String!
-}
-
-type Scene {
-  environment: Environment
-}
-
-type World {
-  id: Int!
-  scene: Scene
-}
